Working version for beam widths of 1, but gets stuck with larger beam widths.

+++ eat/beam_algorithm/beam.py
@@ -47,12 +47,19 @@ class Beam():
 
 class Node():
 
-    def __init__(self, term, array, parent, level):
+    def __init__(self, term, array, parent, level, term_operation):
         self.term = term
+        self.to = term_operation
         self.array = array
         self.parent_node = parent
         self.level = level
         self.proc_hash = None
+
+    def get_fitness(self):
+        """
+        Returns the number of possible solutions for the target array
+        """
+        return self.to.calucate_number_pos_sol(self.array)
 
     def recurse(self):
         node = copy(self)
@@ -79,6 +86,14 @@ class BeamProcessManager():
     def get_processes(self):
         return self.proc_map.values()
 
+    def get_processes_by_decr_fitness(self, level):
+        processes = self.get_process_below_level(level)
+        processes = sorted(processes,
+                           key=lambda bp: bp.node.get_fitness(),
+                           reverse=True)
+        return processes
+
+
     def get_process_below_level(self, level):
         processes_below_h = []
         for bp in self.get_processes():
@@ -97,9 +112,9 @@ class BeamProcessManager():
         for bp in self.get_processes():
             bp.terminate()
 
-    def reset_all(self):
+    def deactivate_all(self):
         for bp in self.get_processes():
-            bp.reset()
+            bp.deactivate()
 
 
 class BeamProcess():
@@ -111,6 +126,7 @@ class BeamProcess():
         self.node = None
 
     def run(self, target, queue, node, **kwargs):
+        node.proc_hash = self.hash
         self.node = node
         self.proc = mp.Process(target=target,
                                args=(queue, node),
@@ -118,9 +134,10 @@ class BeamProcess():
         self.proc.start()
 
     def terminate(self):
-        self.proc.terminate()
+        if self.proc:
+            self.proc.terminate()
 
-    def reset(self):
+    def deactivate(self):
         if self.proc:
             self.proc.terminate()
         self.child_nodes = []
@@ -128,13 +145,16 @@ class BeamProcess():
     def is_alive(self):
         return self.proc.is_alive()
 
+    def __eq__(self, obj):
+        return obj.hash == self.hash
+
 
 class BeamEnumerationAlgorithm():
 
     def __init__(self, groupoid, term_operation, min_term_length=None,
                  max_term_length=None, term_expansion_probability=0.3,
                  male_term_generation_method="random", beam_width=3,
-                 promotion_child_count=2, lr_level_count=5):
+                 lr_level_count=5):
         self.grp = groupoid
         self.to = term_operation
         self.beam = Beam(beam_width)
@@ -147,7 +167,6 @@ class BeamEnumerationAlgorithm():
         self.max_term_length = max_term_length
         self.male_term_generation_method = male_term_generation_method
         self.beam_width = beam_width
-        self.promotion_child_count = promotion_child_count
         self.lr_level_count = lr_level_count
         try:
             mp.set_start_method('fork', force=True)
@@ -197,7 +216,7 @@ class BeamEnumerationAlgorithm():
                                                curr_fnode.array)
             if has_validity_array:
                 return Node(new_female_term, validity_array, curr_fnode,
-                            curr_fnode.level+1)
+                            curr_fnode.level+1, curr_fnode.to)
         else:
             # we couldn't find a valid female term
             return None
@@ -260,7 +279,6 @@ class BeamEnumerationAlgorithm():
                  term_expansion_probability=self.term_expansion_probability,
                  male_term_generation_method=self.male_term_generation_method,
                  beam_width=self.beam_width,
-                 promotion_child_count=self.promotion_child_count,
                  lr_level_count=0)
         # set lr_level_count=0 as to not recurse when finding
         # la/ra sol
@@ -278,7 +296,8 @@ class BeamEnumerationAlgorithm():
         f_node = Node(new_female_term,
                       validity_array,
                       curr_fnode,
-                      curr_fnode.level+1)
+                      curr_fnode.level+1,
+                      curr_fnode.to)
         if f_node:
             if f_node.term not in exclude:
                 mp_queue.put_nowait(f_node)
@@ -289,14 +308,15 @@ class BeamEnumerationAlgorithm():
             male_term_sol = self.to.compute(mt)
             if(self.to.is_solution(male_term_sol, curr_fnode.array)):
                 # found a solution
-                return Node(mt, male_term_sol, curr_fnode, curr_fnode.level+1)
+                return Node(mt, male_term_sol, curr_fnode, curr_fnode.level+1,
+                            curr_fnode.to)
 
     def _print_verbose_valid_term_log(self, f_node_sol,
                                       include_validity_array):
         if self.beam_width > 1:
             f_node_sol_level = [bn.term for bn in
                                 self.beam.get_level(f_node_sol.level)]
-            print("{}: Found {}valid term {} at level {} {}"
+            print("{}: Found {}valid term {} of fitness {} at level {} {}"
                   .format(
                     f_node_sol.parent_node.proc_hash if
                     f_node_sol.parent_node.proc_hash else "Main",
@@ -304,6 +324,7 @@ class BeamEnumerationAlgorithm():
                         if f_node_sol.term in f_node_sol_level
                         else ""),
                     f_node_sol.term,
+                    f_node_sol.get_fitness(),
                     f_node_sol.level,
                     ("with array {}"
                         .format(condensed_array(f_node_sol.array,
@@ -311,28 +332,24 @@ class BeamEnumerationAlgorithm():
                         if include_validity_array else "")
                     ))
         else:
-            print("{}: Found valid term {} at level {} {}"
+            print("{}: Found valid term {} of fitness {} at level {} {}"
                   .format(
                     f_node_sol.parent_node.proc_hash if
                     f_node_sol.parent_node.proc_hash else "Main",
                     f_node_sol.term,
+                    f_node_sol.get_fitness(),
                     f_node_sol.level,
                     ("with array {}"
                         .format(condensed_array(f_node_sol.array,
                                                 self.grp.size))
                         if include_validity_array else "")
                   ))
-            print(f_node_sol.level,
-                  [f_node_sol.term],
-                  condensed_array(f_node_sol.array,
-                                  self.grp.size)
-                  if include_validity_array else "")
 
     def run(self, verbose=False, print_summary=False,
             include_validity_array=False):
         sol_node = None
         # initialize beam at level 0
-        f_node = Node("F", self.to.target, None, 0)
+        f_node = Node("F", self.to.target, None, 0, self.to)
         self.beam.add_level([f_node for _ in range(0, self.beam_width)])
 
         start = time.time()
@@ -362,145 +379,102 @@ class BeamEnumerationAlgorithm():
         while(True):
             try:
                 f_node_sol = mp_queue.get_nowait()
-                # If we reach here we found a valid female term
+                f_node_sol_fitness = f_node_sol.get_fitness()
+                f_node_sol_parent_proc = \
+                    bpm.get_process(f_node_sol.parent_node.proc_hash)
+
+                # If we reach here we found a valid female term of
+                # sufficient fitness
                 if not self.beam.get_level(f_node_sol.level):
                     # Grow the beam
                     self.beam.add_level()
-
-                if (verbose):
+                
+                if verbose:
                     self._print_verbose_valid_term_log(f_node_sol,
                                                        include_validity_array)
 
                 sol_node = self.check_if_has_male_term_solution(f_node_sol)
                 if sol_node:
                     break
+
+                procs_by_fitness = bpm.get_processes_by_decr_fitness(
+                    f_node_sol.level)
+                least_fit_bp = procs_by_fitness[-1]
+                least_fit_bp_fitness = least_fit_bp.node.get_fitness()
+
+                if (f_node_sol_fitness < least_fit_bp_fitness):
+                    # Rerun process for solution node's parent
+                    # and discard f_node_sol
+                    if verbose:
+                        print(f"{f_node_sol_parent_proc.hash}: Discard lower "
+                              f"fitness term {f_node_sol.term} and rerun "
+                              f"{f_node_sol_parent_proc.hash} to search for "
+                              f"a new term")
+                    if f_node_sol.parent_node.level < lrlc:
+                        f_node_sol_parent_proc.run(
+                            self.search_for_valid_female_node_using_lr_array,  # noqa
+                            mp_queue,
+                            f_node_sol.parent_node,
+                            direction=choice(["left", "right"]))
+                    else:
+                        f_node_sol_parent_proc.run(
+                            self.search_for_valid_female_node,
+                            mp_queue,
+                            f_node_sol.parent_node)
                 else:
-                    f_node_sol_proc = \
-                        bpm.get_process(f_node_sol.parent_node.proc_hash)
+                    # add node to beam and parent process child nodes
                     if f_node_sol.term not in \
                             [t.term for t in
-                             self.beam.get_level(f_node_sol.level)]:
+                                self.beam.get_level(f_node_sol.level)]:
                         self.beam.add_node(f_node_sol)
-                        f_node_sol_proc.child_nodes.append(f_node_sol)
-
-                    highest_full_level_num = \
-                        self.beam.get_highest_full_level_number()
-                    # Check if the valid female terms parent has
-                    # produced sufficient children for promotion
-                    if (len(f_node_sol_proc.child_nodes) ==
-                            self.promotion_child_count):
-                        # Terminate productive solution nodes parent
-                        # process and dedicate to a node a level H+1
-                        child_nodes = f_node_sol_proc.child_nodes
-                        child_nodes[0].proc_hash = \
-                            f_node_sol_proc.hash
-                        f_node_sol_proc.reset()
-                        if child_nodes[0].level < lrlc:
-                            bp.run(self.search_for_valid_female_node_using_lr_array,  # noqa
-                                   mp_queue,
-                                   child_nodes[0],
-                                   direction=choice(["left", "right"]))
-                        else:
-                            f_node_sol_proc.run(
-                                self.search_for_valid_female_node,
-                                mp_queue,
-                                child_nodes[0])
-                        # Terminate an unproductive process and
-                        # dedicate to a node at level H+1
-                        least_productive_processes = []
-                        if len(child_nodes) > 1:
-                            least_productive_processes = \
-                                bpm.get_process_below_level(f_node_sol.level)
-                            reassigned_proc_count = 0
-                            if least_productive_processes:
-                                for unproductive_process in \
-                                        least_productive_processes:
-                                    next_child = child_nodes[
-                                        reassigned_proc_count+1]
-                                    next_child.proc_hash = \
-                                        unproductive_process.hash
-                                    unproductive_process.reset()
-                                    unproductive_process.run(
-                                        self.search_for_valid_female_node,
-                                        mp_queue,
-                                        next_child)
-                                    reassigned_proc_count += 1
-                                    if reassigned_proc_count == \
-                                            self.promotion_child_count - 1:
-                                        break
+                        f_node_sol_parent_proc.child_nodes.append(f_node_sol)
+                        
+                        # deactivate the process since it's being reassigned
+                        least_fit_bp.deactivate()
+                        # reassign least fit beam process to work on newly found
+                        # female term
                         if verbose:
-                            promoted_procs = [f_node_sol_proc] + \
-                                least_productive_processes
-                            print("{}-CHILDREN - Promoted {} since {} found "
-                                  "{} children"
-                                  .format(self.promotion_child_count,
-                                          ["{}_{}"
-                                           .format(bp.hash, bp.node.level)
-                                           for bp in promoted_procs],
-                                          f_node_sol_proc.hash,
-                                          self.promotion_child_count))
-                    elif (bpm.get_lowest_process_level() <
-                            highest_full_level_num):
-                        # Check if the beam is full at a level above
-                        # the lowest working process
-                        lowest_processes = bpm.get_process_below_level(
-                            highest_full_level_num)
-                        full_level_f_nodes = self.beam.get_level(
-                            highest_full_level_num)
-                        # kill all lower level processes and move to the
-                        # highest full level
-                        for bp in lowest_processes:
-                            for f_node in full_level_f_nodes:
-                                f_node.proc_hash = bp.hash
-                                bp.reset()
-                                if f_node.level < lrlc:
-                                    bp.run(
-                                        self.search_for_valid_female_node_using_lr_array,  # noqa
-                                        mp_queue,
-                                        f_node,
-                                        direction=choice(["left", "right"]))
-                                else:
-                                    f_node_sol_proc.run(
-                                        self.search_for_valid_female_node,
-                                        mp_queue,
-                                        f_node)
-                        if verbose:
-                            print("FULL-ROW-{} - Promoted {} to level {} "
-                                  "since beam level {} was filled"
-                                  .format(highest_full_level_num,
-                                          [bp.hash for bp in lowest_processes],
-                                          highest_full_level_num,
-                                          highest_full_level_num))
-                    else:
-                        # Create a new process dedicated to the
-                        # solution node's parent
-                        if f_node_sol.parent_node.level < lrlc:
-                            f_node_sol_proc.run(
+                            print(f"{least_fit_bp.hash}: Reassigned "
+                                  f"{least_fit_bp.hash} to work on the newly "
+                                  f"found female term {f_node_sol.term}")
+                        if f_node_sol.level < lrlc:
+                            least_fit_bp.run(
                                 self.search_for_valid_female_node_using_lr_array,  # noqa
                                 mp_queue,
-                                f_node_sol.parent_node,
+                                f_node_sol,
                                 direction=choice(["left", "right"]))
                         else:
-                            f_node_sol_proc.run(
+                            least_fit_bp.run(
                                 self.search_for_valid_female_node,
                                 mp_queue,
-                                f_node_sol.parent_node)
-                        f_node_sol.parent_node.proc = bp
-                        if verbose:
-                            print("CONTINUE - Continue searching with {} for "
-                                  "another child at {}"
-                                  .format(bp.hash, f_node_sol.level))
-
+                                f_node_sol)
+                        if f_node_sol_parent_proc.hash != least_fit_bp.hash:
+                            if verbose:
+                                print(f"{f_node_sol_parent_proc.hash}: Rerun "
+                                      f"{f_node_sol_parent_proc.hash} to "
+                                      f"search for a new child female term "
+                                      f"different than {f_node_sol.term}")
+                            if f_node_sol.parent_node.level < lrlc:
+                                f_node_sol_parent_proc.run(
+                                    self.search_for_valid_female_node_using_lr_array,  # noqa
+                                    mp_queue,
+                                    f_node_sol.parent_node,
+                                    direction=choice(["left", "right"]))
+                            else:
+                                f_node_sol_parent_proc.run(
+                                    self.search_for_valid_female_node,
+                                    mp_queue,
+                                    f_node_sol.parent_node)
                 if verbose:
-                    print("(pid,lvl,chldn): {}".format(
-                          [(bp.hash, bp.node.level,
-                            len(bp.child_nodes))
-                           for bp in bpm.get_processes()]))
+                    print("(pid,lvl,chldn,is_alive): {}".format(
+                        [(bp.hash, bp.node.level,
+                            len(bp.child_nodes), bp.is_alive())
+                        for bp in bpm.get_processes()]))
             except queue.Empty:
                 pass
 
         # kill any remaining processes
-        bpm.reset_all()
+        bpm.deactivate_all()
         mp_queue.close()
 
         node = sol_node.recurse()
@@ -510,5 +484,6 @@ class BeamEnumerationAlgorithm():
         if (print_summary or verbose):
             print_search_summary(node.term, self.to, self.grp, end - start)
         else:
-            print(node.term)
+            if verbose:
+                print(f"Found a solution term {node.term}")
         return node
